requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
anthropic==0.40.0
twilio==9.3.7
httpx==0.27.2
gitpython==3.1.43
